Refine MCP explorations, hosting toggle, and WordPress.com defaults page
- Hosting dashboard toggle: stay on page when opting out, show persistent
  survey notice with "Return to previous version" button
- Global explorations helper: extract variation menu to root-level component
  so it persists across all dashboard pages
- Option 3 (Flat): show Read/Write/Manage links directly on AI and MCP page,
  redirect away from intermediate tools index, fix breadcrumbs to skip
  "MCP access" segment via new excludeHrefs prop on Breadcrumbs
- MCP tools category: adopt Extras toggle card pattern with always-visible
  bold "Enable all" / "Disable all" master toggle per section
- Rename "Default landing page" to "WordPress.com defaults", show primary
  site card unconditionally, move Save button below cards, add primary site
  name badge on preferences summary, switch to WordPress icon
