+++ client/dashboard/me/mcp/tools/category.tsx
@@ -1,14 +1,16 @@
 import { userSettingsQuery, userSettingsMutation } from '@automattic/api-queries';
 import { useSuspenseQuery, useMutation } from '@tanstack/react-query';
 import { __experimentalVStack as VStack, ToggleControl } from '@wordpress/components';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { getAccountMcpAbilities } from '../../../../me/mcp/utils';
 import Breadcrumbs from '../../../app/breadcrumbs';
+import { EXPLORATIONS_STORAGE_KEY } from '../../../app/explorations-helper';
 import { mcpToolsCategoryRoute } from '../../../app/router/me';
 import { Card, CardBody } from '../../../components/card';
 import { PageHeader } from '../../../components/page-header';
 import PageLayout from '../../../components/page-layout';
 import { SectionHeader } from '../../../components/section-header';
+import { Text } from '../../../components/text';
 import {
 	CATEGORY_ORDER,
 	getDisplayCategory,
@@ -25,9 +27,6 @@ interface McpAbility {
 	type?: string;
 	annotations?: { readonly?: boolean; destructive?: boolean };
 }
-
-/** Threshold: sections with more than this many toggles get a master toggle */
-const MASTER_TOGGLE_THRESHOLD = 3;
 
 export default function McpToolsCategory() {
 	const { categorySlug } = mcpToolsCategoryRoute.useParams();
@@ -79,48 +78,53 @@ export default function McpToolsCategory() {
 	const title = permissionLevel?.label ?? __( 'MCP access' );
 	const description = permissionLevel?.description ?? '';
 
+	// Option 3 (Flat) links directly here, skipping the tools index page.
+	// Exclude the "MCP access" breadcrumb segment so the trail reads
+	// "Preferences / AI and MCP / [Read|Write|Manage]".
+	const variation = localStorage.getItem( EXPLORATIONS_STORAGE_KEY );
+	const isFlat = variation === 'C';
+	const excludeHrefs = isFlat ? [ '/me/preferences/ai-and-mcp/tools' ] : undefined;
+	const breadcrumbLength = isFlat ? 3 : 4;
+
 	return (
 		<PageLayout
 			size="small"
 			header={
 				<PageHeader
-					prefix={ <Breadcrumbs length={ 4 } /> }
+					prefix={ <Breadcrumbs length={ breadcrumbLength } excludeHrefs={ excludeHrefs } /> }
 					title={ title }
 					description={ description }
 				/>
 			}
 		>
-			<VStack spacing={ 6 }>
+			<VStack spacing={ 8 }>
 				{ CATEGORY_ORDER.map( ( categoryName ) => {
 					const categoryTools = grouped[ categoryName ];
 					if ( ! categoryTools || categoryTools.length === 0 ) {
 						return null;
 					}
 
-					const allEnabled = categoryTools.every( ( [ , tool ] ) => tool.enabled );
-					const showMasterToggle = categoryTools.length > MASTER_TOGGLE_THRESHOLD;
+					const anyEnabled = categoryTools.some( ( [ , tool ] ) => tool.enabled );
 
 					return (
 						<Card key={ categoryName }>
 							<CardBody>
 								<VStack spacing={ 8 }>
-									<VStack spacing={ 4 }>
-										<SectionHeader level={ 3 } title={ categoryName } />
-										{ showMasterToggle && (
-											<ToggleControl
-												__nextHasNoMarginBottom
-												checked={ allEnabled }
-												disabled={ mutation.isPending }
-												label={ sprintf(
-													/* translators: %s is category name */
-													__( 'Enable all for %s' ),
-													categoryName
-												) }
-												onChange={ ( checked ) => handleSectionToggleAll( categoryTools, checked ) }
-											/>
-										) }
-									</VStack>
-									<VStack spacing={ 4 }>
+									<SectionHeader level={ 3 } title={ categoryName } />
+
+									<ToggleControl
+										__nextHasNoMarginBottom
+										checked={ anyEnabled }
+										disabled={ mutation.isPending }
+										label={
+											<Text weight="bold">
+												{ anyEnabled ? __( 'Disable all' ) : __( 'Enable all' ) }
+											</Text>
+										}
+										onChange={ ( checked ) => handleSectionToggleAll( categoryTools, checked ) }
+									/>
+
+									<VStack>
 										{ categoryTools.map( ( [ toolId, tool ]: [ string, McpAbility ] ) => (
 											<ToggleControl
 												key={ toolId }
