+++ client/dashboard/me/preferences-default-landing/summary.tsx
@@ -1,9 +1,11 @@
-import { rawUserPreferencesQuery } from '@automattic/api-queries';
-import { useSuspenseQuery } from '@tanstack/react-query';
+import { rawUserPreferencesQuery, userSettingsQuery } from '@automattic/api-queries';
+import { useSuspenseQuery, useQuery } from '@tanstack/react-query';
 import { Icon } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import { pin } from '@wordpress/icons';
+import { wordpress } from '@wordpress/icons';
+import { useAppContext } from '../../app/context';
 import RouterLinkSummaryButton from '../../components/router-link-summary-button';
+import { getSiteDisplayName } from '../../utils/site-name';
 import type { SummaryButtonBadgeProps } from '@automattic/components/src/summary-button/types';
 
 type LandingPage = 'primary-site-dashboard' | 'sites' | 'reader';
@@ -15,6 +17,8 @@ const LANDING_PAGE_LABELS: Record< LandingPage, string > = {
 };
 
 export default function PreferencesDefaultLandingSummary() {
+	const { queries } = useAppContext();
+
 	const { data: landingPage } = useSuspenseQuery( {
 		...rawUserPreferencesQuery(),
 		select: ( preferences ): LandingPage => {
@@ -28,19 +32,39 @@ export default function PreferencesDefaultLandingSummary() {
 		},
 	} );
 
+	const { data: primarySiteId } = useSuspenseQuery( {
+		...userSettingsQuery(),
+		select: ( data ) => data.primary_site_ID,
+	} );
+
+	const { data: sites } = useQuery(
+		queries.sitesQuery( { site_visibility: 'visible', include_a8c_owned: false } )
+	);
+
+	const primarySite = sites?.find( ( site ) => site.ID === primarySiteId );
+	const primarySiteName = primarySite ? getSiteDisplayName( primarySite ) : undefined;
+
 	const badges: SummaryButtonBadgeProps[] = [
 		{
 			text: LANDING_PAGE_LABELS[ landingPage ],
 			intent: 'default',
 		},
+		...( primarySiteName
+			? [
+					{
+						text: primarySiteName,
+						intent: 'default' as const,
+					},
+			  ]
+			: [] ),
 	];
 
 	return (
 		<RouterLinkSummaryButton
 			to="/me/preferences/default-landing-page"
-			title={ __( 'Default landing page' ) }
-			description={ __( 'Choose what you see after logging into WordPress.com' ) }
-			decoration={ <Icon icon={ pin } /> }
+			title={ __( 'WordPress.com defaults' ) }
+			description={ __( 'Set your starting point after you log in and primary site.' ) }
+			decoration={ <Icon icon={ wordpress } /> }
 			badges={ badges }
 		/>
 	);
